+++ NEWS.rst
@@ -3,6 +3,10 @@ News
 
 Lists of changes between versions.
 
+Latest
+------
+* Minor: news_files, build_steps, lint_steps, test_steps can be set to nothing to disable them.
+
 2.4.0
 -----
 * Minor: Improved cli input handling and help description.